Expand the example to upload vertices
This uploads data (positions) to opengl via a vertex buffer object
and vertex array object.  It uses unsafe transmutation code which
could be replaced by an external crate, but does it manually to
prevent dependency bloat (and to express exactly what is going on).

Further, the GLSL version number for SDL2 is downgraded to 130 so
as to run on the OpenGL3.0 requested from SDL2.